requirements.txt

django-mp-config
